requirements.txt
cycler==0.10.0
decorator==4.4.1
imageio==2.6.1
imutils==0.5.3
kiwisolver==1.1.0
lxml==4.4.2
matplotlib==3.1.2
MouseInfo==0.1.2
networkx==2.4
numpy==1.17.4
opencv-python==4.1.2.30
Pillow==6.2.1
PyAutoGUI==0.9.48
PyGetWindow==0.0.8
PyMsgBox==1.0.7
pyparsing==2.4.5
pyperclip==1.7.0
PyRect==0.1.4
PyScreeze==0.1.25
python-dateutil==2.8.1
python-docx==0.8.10
PyTweening==1.0.3
PyWavelets==1.1.1
scikit-image==0.16.2
scipy==1.3.3
six==1.13.0
win32core==221.36
